Sorting by clicking at column in Krokiet, workaround for low performance of TextEdit (#1718)

+++ czkawka_core/src/tools/same_music/mod.rs
@@ -84,7 +84,7 @@ struct GroupedFilesToCheck {
 #[derive(Default, Clone)]
 pub struct Info {
     pub number_of_duplicates: usize,
-    pub number_of_groups: u64,
+    pub number_of_groups: usize,
     pub scanning_time: Duration,
 }
